pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.10.0",
    "fastapi>=0.133.0",
    "groq>=1.0.0",
    "nltk>=3.9.3",
    "pinecone>=8.1.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.7.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]
